+++ src/output/worldStateManager/WorldStateManager.cpp
@@ -566,6 +566,21 @@ FunctionHook<void> onCountSetItemsMaybe(0x0046BD20, []()-> void
     AddSetToLevel(WARP_FROM_PAST, LevelAndActIDs_Past2, Characters_Big);
 });
 
+
+// We move the emblem a little higher so Cream can get it.
+FunctionHook<void, task*> onEmblemMain(0x4B4940, [](task* tp)-> void
+{
+    if (levelact(CurrentLevel, CurrentAct) == LevelAndActIDs_StationSquare1)
+        if (tp->twp->pos.x > 386 && tp->twp->pos.x < 390
+            && tp->twp->pos.y > -6 && tp->twp->pos.y < -5
+            && tp->twp->pos.z > 489 && tp->twp->pos.z < 492)
+        {
+            tp->twp->pos.y = -4;
+        }
+
+    onEmblemMain.Original(tp);
+});
+
 FunctionHook<void> onMissionSetLoad(0x591A70, []()-> void
 {
     if (!worldStateManagerPtr->options.missionModeEnabled)