+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - Improved emblem requirements calculation.
     - It should hopefully fix the issue with the wrong emblem percentage goal.
 - Fixed soft lock when losing any race as Tails.
+- Moved the City Hall emblem a little higher so you can get it when using the Cream the Rabbit mod.
 
 ## [v0.7.1] - 2024-08-22